Fix tainted string error when updating comment
When updating a comment, a tainted string error was being thrown due to
the lack of a h() escape on the article title in the comment partial.
A deprecation warning from truncate() in the comment partial has also
been resolved.

A functional test (test_should_update_comment) for the admin comments
controller is also included.